// navplan_frontend/angular-app/src/app/flight-map/state/ngrx/map-layer-type.ts
export enum MapLayerType {
    AIRPORTS_NAVAIDS = 'airportsNavaids',
    REPORTING = 'reporting',
    CIRCUITS = 'circuits',
    AIRSPACES = 'airspaces',
    AIRSPACE_CAT = 'airspaceCat',
    AIRSPACE_GLIDER = 'airspaceGlider',
    AIRSPACE_RESTRICTED = 'airspaceRestricted',
    NOTAMS = 'notams',
    WEBCAMS = 'webcams'
}